feat: centralize and co-locate types for auth utility functions
